imported project 2
The multi-modal analysis known as CITE-seq, a single-cell sequencing assay, allows for the relationship between ribonucleic acid (RNA) expression and protein expression to be explored by measuring both simultaneously in single cells. Using CITE-seq, protein expression is measured using Antibody-Derived Tags (ADT) and RNA is quantified using single-cell RNA sequencing. The ability to predict what proteins will be expressed given gene expression has been a goal of the scientific community for decades. Thus, the goal of this project will be to design an algorithm to predict the protein expression profile of 25 proteins in 1000 test cells given the gene expression of 639 genes in those cells. The Ordinary Least Squares (OLS) regression method will be implemented to determine the linear regression equation which describes the relationship between protein and RNA gene expression.